# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReelStatus = typeof ReelStatus[keyof typeof ReelStatus];


export const ReelStatus = {
  published: 'published',
  draft: 'draft',
} as const;

export interface Reel {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  videoPath: string;
  /** @nullable */
  thumbnailPath?: string | null;
  status: ReelStatus;
  views: number;
  likes: number;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ReelInputStatus = typeof ReelInputStatus[keyof typeof ReelInputStatus];


export const ReelInputStatus = {
  published: 'published',
  draft: 'draft',
} as const;

export interface ReelInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  videoPath: string;
  thumbnailPath?: string;
  status?: ReelInputStatus;
  duration?: number;
}

export type ReelUpdateStatus = typeof ReelUpdateStatus[keyof typeof ReelUpdateStatus];


export const ReelUpdateStatus = {
  published: 'published',
  draft: 'draft',
} as const;

export interface ReelUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailPath?: string;
  status?: ReelUpdateStatus;
  duration?: number;
}

export interface ReelListResponse {
  reels: Reel[];
  total: number;
  page: number;
  limit: number;
}

export interface ReelsStats {
  totalReels: number;
  totalViews: number;
  totalLikes: number;
  publishedCount: number;
  draftCount: number;
  recentUploads: Reel[];
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListReelsParams = {
page?: number;
limit?: number;
status?: ListReelsStatus;
};

export type ListReelsStatus = typeof ListReelsStatus[keyof typeof ListReelsStatus];


export const ListReelsStatus = {
  published: 'published',
  draft: 'draft',
  all: 'all',
} as const;