# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reels Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reels
    description: Reel video operations
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reels:
    get:
      operationId: listReels
      tags: [reels]
      summary: List all reels (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [published, draft, all]
            default: published
      responses:
        "200":
          description: List of reels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReelListResponse"
    post:
      operationId: createReel
      tags: [reels]
      summary: Create a new reel entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReelInput"
      responses:
        "201":
          description: Reel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"

  /reels/stats:
    get:
      operationId: getReelsStats
      tags: [reels]
      summary: Get aggregate stats for admin dashboard
      responses:
        "200":
          description: Aggregate statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReelsStats"

  /reels/{id}:
    get:
      operationId: getReel
      tags: [reels]
      summary: Get a single reel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reel found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
        "404":
          description: Not found
    patch:
      operationId: updateReel
      tags: [reels]
      summary: Update a reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
        "404":
          description: Not found
    delete:
      operationId: deleteReel
      tags: [reels]
      summary: Delete a reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /reels/{id}/view:
    post:
      operationId: incrementReelView
      tags: [reels]
      summary: Increment view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: View incremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"

  /reels/{id}/like:
    post:
      operationId: toggleReelLike
      tags: [reels]
      summary: Toggle like on a reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Reel:
      type: object
      required: [id, title, videoPath, status, views, likes, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        videoPath:
          type: string
        thumbnailPath:
          type: ["string", "null"]
        status:
          type: string
          enum: [published, draft]
        views:
          type: integer
        likes:
          type: integer
        duration:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReelInput:
      type: object
      required: [title, videoPath]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        videoPath:
          type: string
        thumbnailPath:
          type: string
        status:
          type: string
          enum: [published, draft]
          default: published
        duration:
          type: number

    ReelUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        thumbnailPath:
          type: string
        status:
          type: string
          enum: [published, draft]
        duration:
          type: number

    ReelListResponse:
      type: object
      required: [reels, total, page, limit]
      properties:
        reels:
          type: array
          items:
            $ref: "#/components/schemas/Reel"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    ReelsStats:
      type: object
      required: [totalReels, totalViews, totalLikes, publishedCount, draftCount, recentUploads]
      properties:
        totalReels:
          type: integer
        totalViews:
          type: integer
        totalLikes:
          type: integer
        publishedCount:
          type: integer
        draftCount:
          type: integer
        recentUploads:
          type: array
          items:
            $ref: "#/components/schemas/Reel"

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string