# lib/api-zod/src/generated/types/reel.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReelStatus } from './reelStatus';

export interface Reel {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  videoPath: string;
  /** @nullable */
  thumbnailPath?: string | null;
  status: ReelStatus;
  views: number;
  likes: number;
  /** @nullable */
  duration?: number | null;
  createdAt: Date;
  updatedAt: Date;
}