# lib/api-zod/src/generated/types/reelInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReelInputStatus } from './reelInputStatus';

export interface ReelInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  videoPath: string;
  thumbnailPath?: string;
  status?: ReelInputStatus;
  duration?: number;
}