# lib/api-zod/src/generated/types/reelListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Reel } from './reel';

export interface ReelListResponse {
  reels: Reel[];
  total: number;
  page: number;
  limit: number;
}