# lib/api-zod/src/generated/types/reelUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReelUpdateStatus } from './reelUpdateStatus';

export interface ReelUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailPath?: string;
  status?: ReelUpdateStatus;
  duration?: number;
}