# lib/api-zod/src/generated/types/reelsStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Reel } from './reel';

export interface ReelsStats {
  totalReels: number;
  totalViews: number;
  totalLikes: number;
  publishedCount: number;
  draftCount: number;
  recentUploads: Reel[];
}