# lib/api-zod/src/generated/types/uploadUrlRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Reels Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}